Add files via upload
In this project, goal is to develop a sentiment analysis recommendation system to identify emotion and thoughts of cancer patients through the reviews or feedback input data.

Dataset : Emotional responses are recorded and dataset is prepared by the Mendeley team - https://data.mendeley.com/datasets/69dcnv2gzd/1 . (Mental Health Insights: Vulnerable Cancer Survivors & Caregivers Data)

 The dataset was collected from various platforms such as Reddit, Daily Strength, and the Health Board, comprising a total of over 10,087 posts. The posts are related to five types of cancer: brain, colon, liver, leukemia, and lung cancer.
The dataset was analyzed by two team members who scored each post based on the emotions expressed, utilizing a scale from -2 to 1. The scoring system is designed to capture a spectrum of emotions. Negative scores (-1 or -2) were assigned to posts expressing grief or suffering, positive scores (1) were given for posts with happy emotions like relief or accomplishment, and posts with no discernible emotion received a score of 0, categorizing them as neutral.

Prediction :   Aim is achieved by preparing an alogrithm which anlyses the text response from the user and provides feedback to be used by care givers, doctores etc for treatment of the patients. The primary objective of this analysis is to understand the emotional aspects conveyed in the posts of cancer patients.